Update view-locator.md
I created an Avalonia app using the wizard project template, and it didn't have a ViewLocator generated for it or referenced in its App.axaml file. Having an explanation for where it's used seems appropriate, so that clients don't need to spend as much time as I did trying to get it working.