Improve example in accessing-the-ui-thread.md
I think the current example is not the best fit to demonstrate how to call the ui thread from another thread. Long running work can simply be solved with await Task.Run(() =>{...}).

Also I suspect that the current example description is not correct. It says "Notice that because the long running task is executed on its own thread, the UI does not lose responsiveness."
But the LongRunningTask() Function is actually running on the UIThread. It's only not blocking the ui thread because it uses Task.Delay instead of Thread.Sleep.

Thats why I created a new sample with a usecase I find more fitting. I hope that it is helpfull.