Improve accessing-the-ui-thread.md example
Change the example from MainViewModel to MainView and change the Text of a TextBlock instead of changing a ViewModel property called Text. The reason is that changing the Text property from another thread seems to be okay, no System.InvalidOperationException is thrown. With the new code which changes the TextBlock text directly the exception is thrown.

Add a commented out example code which shows how to provoke the System.InvalidOperationException by calling SetText from outside the ui thread.

Add a try-catch to the async void method. It is a good practice because exceptions thrown inside it can't be catched by the caller.

Simplify the names of SetUiText and GetUiText to SetText and GetText and use Expressions to save some code lines.