Update compiled-bindings.md to rewrite section about type casting
Type casting was no longer needed as of 11.3.0. The docs continued to describe type casting, which caused user confusion, not helped by Rider highlighting the inference as an error. This rewrite aims to clarify how DataContext inference works, which version it was introduced in, and when users need to to explicit type casting.